Add a fake ping clamping feature: g_ping_handicap
It doesn't actually change someone's ping, just the
number shown in the game. The idea is you set this to
a low but reasonble value like 25 and players with a
lower ping than this threshold will display the threshold
value instead.

This is to combat league matches where servers are refused
because one or more players pings too low and is seen as
having an advantage. It doesn't affect the players at all.

There is a slight amount of jitter added to make it seem
real.